refactor: Line이 Section을 포함하도록 수정
