+++ src/main/java/nextstep/subway/applicaion/StationLineService.java
@@ -4,9 +4,12 @@ import java.util.List;
 import java.util.stream.Collectors;
 import nextstep.subway.applicaion.dto.StationLineRequest;
 import nextstep.subway.applicaion.dto.StationLineResponse;
+import nextstep.subway.applicaion.dto.StationSectionRequest;
+import nextstep.subway.applicaion.dto.StationSectionResponse;
 import nextstep.subway.domain.Station;
 import nextstep.subway.domain.StationLine;
 import nextstep.subway.domain.StationLineRepository;
+import nextstep.subway.domain.StationSection;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -16,13 +19,16 @@ public class StationLineService {
     private final StationLineRepository stationLineRepository;
     private final StationService stationService;
     private final StationLineMapper stationLineMapper;
+    private final StationSectionMapper stationSectionMapper;
 
     public StationLineService(StationLineRepository stationLineRepository,
                               StationService stationService,
-                              StationLineMapper stationLineMapper) {
+                              StationLineMapper stationLineMapper,
+                              StationSectionMapper stationSectionMapper) {
         this.stationLineRepository = stationLineRepository;
         this.stationService = stationService;
         this.stationLineMapper = stationLineMapper;
+        this.stationSectionMapper = stationSectionMapper;
     }
 
     @Transactional
@@ -57,6 +63,27 @@ public class StationLineService {
     @Transactional
     public void deleteStationLine(Long lineId) {
         stationLineRepository.deleteById(lineId);
+    }
+
+    @Transactional
+    public StationSectionResponse createStationSection(StationSectionRequest request, Long lineId) {
+        validateNoneExistStation(request);
+        StationSection stationSection = stationSectionMapper.of(request);
+        StationLine stationLine = findLineById(lineId);
+        stationLine.addSection(stationSection);
+        return stationSectionMapper.of(stationSection);
+    }
+
+    @Transactional
+    public void deleteStationSection(Long lineId, Long stationId) {
+        StationLine stationLine = findLineById(lineId);
+        StationSection section = stationLine.findSectionByDownStationId(stationId);
+        stationLine.deleteSection(section);
+    }
+
+    private void validateNoneExistStation(StationSectionRequest request) {
+        stationService.findStationById(request.getDownStationId());
+        stationService.findStationById(request.getUpStationId());
     }
 
     protected StationLine findLineById(Long lineId) {
